+++ lib/MongoDB/algorithms/store.js
@@ -1,5 +1,5 @@
 const Collection = require('../Collection');
-const Versions = require('./versions');
+const Versions = require('../versions');
 const Builds = require('./builds');
 const Readme = require('./readme');
 const collections = require('../collections');
