+++ lib/MongoDB/versions.js
@@ -1,6 +1,6 @@
-const Collection = require('../Collection');
-const collections = require('../collections');
-const Query = require('../Query');
+const Collection = require('./Collection');
+const collections = require('./collections');
+const Query = require('./Query');
 
 class Versions extends Collection {
     constructor(db, client) {
